Hospital_Managment/requirements.txt
django==6.0.4
djangorestframework
Pillow
twilio
python-dotenv
